Rename Vue files
