+++ src/main/webapp/app/common/primary/app/App.component.ts
