+++ src/main/webapp/app/common/primary/home/Home.component.ts
@@ -1,0 +1,3 @@
+import { Vue } from 'vue-class-component';
+
+export default class HomeComponent extends Vue {}
