--- src/main/webapp/app/common/primary/home/home.component.ts
@@ -1,3 +1,0 @@
-import { Vue } from 'vue-class-component';
-
-export default class Home extends Vue {}
