+++ src/main/webapp/app/main.ts
@@ -1,5 +1,5 @@
 import { createApp } from 'vue';
-import AppVue from './common/primary/app/app.vue';
+import AppVue from './common/primary/app/App.vue';
 import router from './router';
 
 createApp(AppVue).use(router).mount('#app');
