+++ src/main/webapp/app/router/index.ts
@@ -1,5 +1,5 @@
 import { createRouter, createWebHistory, RouteRecordRaw } from 'vue-router';
-import Home from '../common/primary/home/home.vue';
+import Home from '../common/primary/home/Home.vue';
 
 const routes: Array<RouteRecordRaw> = [
   {
@@ -13,7 +13,7 @@ const routes: Array<RouteRecordRaw> = [
     // route level code-splitting
     // this generates a separate chunk (about.[hash].js) for this route
     // which is lazy-loaded when the route is visited.
-    component: () => import(/* webpackChunkName: "about" */ '../common/primary/home/about.vue'),
+    component: () => import(/* webpackChunkName: "about" */ '../common/primary/home/About.vue'),
   },
 ];
