+++ src/test/javascript/spec/common/primary/app/App.component.spec.ts
@@ -1,7 +1,7 @@
 import { shallowMount } from '@vue/test-utils';
-import AppVue from '@/common/primary/app/app.vue';
+import AppVue from '@/common/primary/app/App.vue';
 
-describe('app.vue', () => {
+describe('App.vue', () => {
   it('renders props.msg when passed', () => {
     const wrapper = shallowMount(AppVue, {});
     expect(wrapper.exists()).toBe(true);
