+++ src/test/javascript/spec/common/primary/home/Home.component.spec.ts
@@ -1,7 +1,7 @@
 import { shallowMount } from '@vue/test-utils';
-import HomeVue from '@/common/primary/home/home.vue';
+import HomeVue from '@/common/primary/home/Home.vue';
 
-describe('home.vue', () => {
+describe('Home.vue', () => {
   it('renders props.msg when passed', () => {
     const wrapper = shallowMount(HomeVue, {});
     expect(wrapper.exists()).toBe(true);
